Fix qnap openvpn issue by renaming: openvpn -> openvpn-qnapfix
For some strange reason, some versions of /sbin/server_mgr in QOS
4.3.x kill openvpn processes (even if running inside a container).
This might be tied to QVPN (without it, it doesn't seem to happen).
To check if your QOS installation is affected, check for a line
like:

  DAEMON## = openvpn, stop, /usr/sbin/openvpn

in /etc/daemon_mgr.conf. But to avoid the issue altogether, this
commit renames the openvpn binary to openvpn-qnapfix.